Move admin api impl to its own package
It doesn't really belong under rest/client/v1 any more.
